Clean the code up and add documentation comments
