+++ .gitignore
@@ -1,5 +1,6 @@
 # My:
 /priv
+/database
 # ======================================
 ### Windows template:
 # Windows thumbnail cache files
