+++ public/js/messSys.js
@@ -1,38 +1,9 @@
 import { getNowTime } from "./time.js";
 
-export function clearChat() {
-  const chat = document.getElementById("chat");
-  const messages = chat.querySelectorAll("div.message");
-  messages.forEach(message => {
-    message.remove();
-  });
-}
-
-
-export function loadChatHeader(nickname, active, profilePhoto) {
-  const header = document.getElementById("chat-header");
-
-  // Change name
-  const nicknameP = header.querySelector("p.nickname");
-  nicknameP.innerText = nickname;
-
-  // Change status
-  const statusDivs = header.querySelectorAll("div.status");
-  if (active) {
-    statusDivs[0].classList.add("active");
-    statusDivs[1].classList.remove("active");
-  } else {
-    statusDivs[1].classList.add("active");
-    statusDivs[0].classList.remove("active");
-  }
-
-  // Change profile photo:
-  const photoImg = header.querySelector("div.profile-img img.profile-photo");
-  photoImg.src = profilePhoto;
-}
-
-
-export function addNewMessage(loadMode = false, amIAuthor, messTxt, isLast = false, sentAt = 0) {
+// =====================================================================================================
+// PRIVATE:
+// =====================================================================================================
+function addNewMessage(loadMode = false, amIAuthor, messTxt, isLast = false, sentAt = 0) {
   const chat = document.getElementById("chat");
 
   const message = document.createElement("p");
@@ -59,7 +30,7 @@ export function addNewMessage(loadMode = false, amIAuthor, messTxt, isLast = fal
   }
   messBubble.appendChild(message);
 
-  
+  // Add messBubble to chat interface (in one of two way)
   if(loadMode) chat.insertBefore(messBubble, chat.firstChild);
   else chat.appendChild(messBubble);
 
@@ -89,6 +60,74 @@ function loadMessages(data) {
 
 }
 
+// =====================================================================================================
+// PUBLIC:
+// =====================================================================================================
+// Remove all message bubbles from chat:
+export function clearChat() {
+  const chat = document.getElementById("chat");
+  const messages = chat.querySelectorAll("div.message");
+  messages.forEach(message => {
+    message.remove();
+  });
+}
+
+
+// Load new chat header:
+export function loadChatHeader(nickname, active, profilePhoto) {
+  const header = document.getElementById("chat-header");
+
+  // Change name
+  const nicknameP = header.querySelector("p.nickname");
+  nicknameP.innerText = nickname;
+
+  // Change status
+  const statusDivs = header.querySelectorAll("div.status");
+  if (active) {
+    statusDivs[0].classList.add("active");
+    statusDivs[1].classList.remove("active");
+  } else {
+    statusDivs[1].classList.add("active");
+    statusDivs[0].classList.remove("active");
+  }
+
+  // Change profile photo:
+  const photoImg = header.querySelector("div.profile-img img.profile-photo");
+  photoImg.src = profilePhoto;
+}
+
+
+// Send new message:
+export function sendNewMessage(event, userID) {
+  event.preventDefault();
+  const authorID = userID;                 // User always is author of new message. Getting userID
+  const sentAt = getNowTime();
+
+  const chatTextArea = document.getElementById("write-mess");
+  let messTxt = chatTextArea.value;
+  chatTextArea.value = "";                 // Clean textarea
+
+  // If it's empty do nothing:
+  if(messTxt === "" || messTxt === "\n" || messTxt === "\r\n") return 0;
+
+  // Add message to interface chat:
+  const itsMyMess = true;                   // TODO: To remove
+  addNewMessage(false, itsMyMess, messTxt);
+
+  // Get inboxID:
+  const chat = document.getElementById("chat");
+  const inboxID = chat.getAttribute('active-inbox');
+
+  // Send new message with data to server:
+  console.log(`User '${authorID}' send new mess: "${messTxt}" to '${inboxID}' inbox at ${sentAt}.`);
+  const url = `/chat/send/${inboxID}/${authorID}/${sentAt}/${messTxt}`;
+  fetch(url, {
+      method: 'POST',
+  })
+}
+
+
+// Load chat conversation with friend:
 export function loadConversation(friendID, lastMessTime = 0) {
   fetch(`/chat/load/${friendID}`, {
     method: "GET",
