+++ public/js/time.js
@@ -14,3 +14,9 @@ export function getNowTime() {
 
   return time;
 }
+
+
+// Delay:
+export function delay(time) {
+  return new Promise(resolve => setTimeout(resolve, time));
+}
