+++ routes/sendMessage.js
@@ -1,5 +1,11 @@
 const { createClientDB } = require("./test_db");
 
+// =====================================================================================================
+// PRIVATE:
+// =====================================================================================================
+/**
+ * Add new message and its data into message table in database.
+ */
 function addNewMessageToDB(inboxID, authorID, sentAt, messTxt) {
   const client = createClientDB();
   client
@@ -14,6 +20,9 @@ function addNewMessageToDB(inboxID, authorID, sentAt, messTxt) {
     .finally(() => client.end());
 }
 
+/**
+ * Update information about last message in inbox table.
+ */
 function updateInboxData(inboxID, authorID, sentAt, messTxt) {
   const client = createClientDB();
   client
@@ -29,7 +38,13 @@ function updateInboxData(inboxID, authorID, sentAt, messTxt) {
     .finally(() => client.end());
 }
 
-
+// =====================================================================================================
+// PUBLIC:
+// =====================================================================================================
+/**
+ * Function gets data from URL parametrs and call function to add new data in DB and
+ * update data in inbox table.
+ */
 function sendMessage(app) {
   app.post("/chat/send/:inboxID/:authorID/:sentAt/:messTxt", (req, res) => {
     const inboxID = req.params.inboxID;
@@ -43,4 +58,6 @@ function sendMessage(app) {
   });
 }
 
+// =====================================================================================================
+// Export:
 module.exports = sendMessage;
